requirements.txt
streamlit
pandas
Pillow
numpy
opencv-python-headless  # Ensure this version works without GUI
pyzbar
pymongo